chore(app): 调整部分代码
