+++ .env
@@ -1,1 +1,1 @@
-VUE_APP_API_URL=http://106.54.69.78:8080/api
+VUE_APP_API_URL=https://huiyou.fun/api
