+++ package.json
@@ -31,7 +31,7 @@
     "eslint-plugin-prettier": "^3.1.4",
     "eslint-plugin-vue": "^7.0.0-0",
     "less": "^3.0.4",
-    "less-loader": "^5.0.0",
+    "less-loader": "^6.0.0",
     "prettier": "^2.1.2",
     "typescript": "~3.9.3"
   }
